Don't use a store abstraction anymore
