+++ src/adapters/store.js
@@ -1,9 +1,9 @@
 import Error from '../adapters/error';
 import { version } from '../../config/constants.yml';
 import { findIndexIgnoreCase } from '../libs/string';
-import LocalStore from '../libs/local_store';
 import { getKey } from 'src/libs/pois';
 import { fire } from 'src/libs/customEvents';
+import { isPoiCompliantKey } from 'src/libs/pois';
 
 export default class Store {
   constructor() {
@@ -13,44 +13,67 @@ export default class Store {
     }
     // if store not initialized, use this
     window.__store = this;
-
-    // init stores
-    this.localStore = new LocalStore();
-    this.abstractStore = this.localStore;
   }
 
   async getAllPois() {
+    let localStorageKeys = [];
     try {
-      return await this.abstractStore.getAllPois();
+      localStorageKeys = Object.keys(localStorage);
     } catch (e) {
-      Error.sendOnce('store', 'getAllPois',
-        `error getting pois from ${this.abstractStore.storeName}`, e);
+      Error.sendOnce('local_store', 'getAllPois', 'error getting pois keys', e);
       return [];
+    }
+    const items = localStorageKeys.reduce((filtered, k) => {
+      if (isPoiCompliantKey(k)) {
+        try {
+          const poi = JSON.parse(localStorage.getItem(k));
+          filtered.push(poi);
+        } catch (e) {
+          Error.sendOnce('local_store', 'getAllPois', 'error getting pois', e);
+        }
+      }
+      return filtered;
+    }, []);
+    return items;
+  }
+
+  async get(k) {
+    try {
+      return JSON.parse(localStorage.getItem(k));
+    } catch (e) {
+      Error.sendOnce('local_store', 'get', `error parsing item with key ${k}`, e);
+      return null;
+    }
+  }
+
+  async set(k, v) {
+    try {
+      localStorage.setItem(k, JSON.stringify(v));
+    } catch (e) {
+      Error.sendOnce('local_store', 'set', 'error setting item', e);
     }
   }
 
   async getLastLocation() {
     try {
-      return await this.abstractStore.get(`qmaps_v${version}_last_location`);
+      return this.get(`qmaps_v${version}_last_location`);
     } catch (e) {
-      Error.sendOnce('store', 'getLastLocation',
-        `error getting last location from ${this.abstractStore.storeName}`, e);
+      Error.sendOnce('store', 'getLastLocation', 'error getting last location', e);
       return null;
     }
   }
 
   async setLastLocation(loc) {
     try {
-      return await this.abstractStore.set(`qmaps_v${version}_last_location`, loc);
+      return this.set(`qmaps_v${version}_last_location`, loc);
     } catch (e) {
-      Error.sendOnce('store', 'setLastLocation',
-        `error setting location in ${this.abstractStore.storeName}`, e);
+      Error.sendOnce('store', 'setLastLocation', 'error setting location', e);
       throw e;
     }
   }
 
   async getMatches(term) {
-    const storedItems = await this.abstractStore.getAllPois();
+    const storedItems = await this.getAllPois();
     return storedItems.filter(storedItem => {
       return findIndexIgnoreCase(storedItem.name, term) !== -1;
     });
@@ -58,36 +81,35 @@ export default class Store {
 
   async has(poi) {
     try {
-      return await this.abstractStore.has(getKey(poi));
+      return Boolean(await this.get(getKey(poi)));
     } catch (e) {
-      Error.sendOnce('store', 'has',
-        `error checking existing key in ${this.abstractStore.storeName}`, e);
+      Error.sendOnce('store', 'has', 'error checking existing key', e);
     }
   }
 
   async add(poi) {
     try {
-      await this.abstractStore.set(getKey(poi), poi);
+      await this.set(getKey(poi), poi);
       fire('poi_added_to_favs', poi);
     } catch (e) {
-      Error.sendOnce('store', 'add', `error adding poi in ${this.abstractStore.storeName}`, e);
+      Error.sendOnce('store', 'add', 'error adding poi', e);
     }
   }
 
   async del(poi) {
     try {
-      await this.abstractStore.del(getKey(poi));
+      localStorage.removeItem(getKey(poi));
       fire('poi_removed_from_favs', poi);
     } catch (e) {
-      Error.sendOnce('store', 'del', `error deleting key from ${this.abstractStore.storeName}`, e);
+      Error.sendOnce('store', 'del', 'error removing item', e);
     }
   }
 
   async clear() {
     try {
-      await this.abstractStore.clear();
+      localStorage.clear();
     } catch (e) {
-      Error.sendOnce('store', 'clear', `error clearing ${this.abstractStore.storeName}`, e);
+      Error.sendOnce('local_store', 'clear', 'error clearing store', e);
     }
   }
 }
